join the InventoryActivity with our previous ui_list_zones
