--- app/src/main/java/ekylibre/database/ui_list_zones.java
@@ -1,71 +1,0 @@
-package ekylibre.database;
-
-import android.app.Dialog;
-import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.widget.LinearLayout;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import ekylibre.zero.R;
-import ekylibre.zero.SelectZoneDialogFragment;
-import ekylibre.zero.UiSelectZoneAdapter;
-
-public class ui_list_zones extends AppCompatActivity implements SelectZoneDialogFragment.OnFragmentInteractionListener {
-    public List<String> list;
-    private SelectZoneDialogFragment fragment;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_ui_list_zones);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-
-        FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show();
-            }
-        });
-
-        list = new ArrayList<>();
-        for (int i=0;i<10;i++){
-            list.add("Item_"+i);
-        }
-
-        fragment = SelectZoneDialogFragment.newInstance(list);
-        fragment.show(getFragmentTransaction(),"dialog");
-
-    }
-
-
-    @Override
-    public void onFragmentInteraction(Object object) {
-
-    }
-
-    private FragmentTransaction getFragmentTransaction (){
-        FragmentTransaction ft = getSupportFragmentManager().beginTransaction();
-        Fragment prev = getSupportFragmentManager().findFragmentByTag("dialog");
-        if (prev!= null){
-            ft.remove(prev);
-        }
-        ft.addToBackStack(null);
-        return ft;
-    }
-}
