+++ app/src/main/java/ekylibre/zero/inventory/SelectZoneDialogFragment.java
@@ -1,31 +1,26 @@
-package ekylibre.zero;
+package ekylibre.zero.inventory;
 
-import android.app.Dialog;
 import android.content.Context;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v4.app.DialogFragment;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.LinearLayout;
 
-import java.util.ArrayList;
 import java.util.List;
+
+import ekylibre.zero.R;
+import ekylibre.zero.inventory.adapters.UiSelectZoneAdapter;
 
 public class SelectZoneDialogFragment extends DialogFragment {
 
     private OnFragmentInteractionListener fragmentListener;
-    public List<String> zoneList;
+    public List<ItemZoneInventory> zoneList;
 
     public SelectZoneDialogFragment(){    }
 
@@ -41,7 +36,7 @@ public class SelectZoneDialogFragment extends DialogFragment {
         }
     }
 
-    public static SelectZoneDialogFragment newInstance(List<String> list){
+    public static SelectZoneDialogFragment newInstance(List<ItemZoneInventory> list){
         SelectZoneDialogFragment fragment = new SelectZoneDialogFragment();
         fragment.zoneList = list;
         return fragment;
